Enabling FormatParser to handle different results.
