+++ lib/document.rb
@@ -1,0 +1,5 @@
+module FormatParser
+  class Document
+    # TODO: Figure out which is the metainformation revelant to Documents.
+  end
+end
