+++ lib/format_parser.rb
@@ -1,20 +1,14 @@
 require 'thread'
 
+Dir.glob(__dir__ + '/**/*.rb').sort.each { |f| require f }
+
 module FormatParser
-  require_relative 'file_information'
-  require_relative 'io_utils'
-  require_relative 'read_limiter'
-  require_relative 'remote_io'
-  require_relative 'care'
-
-  PARSER_MUX = Mutex.new
-
-  def self.register_parser_constructor(object_responding_to_new)
-    PARSER_MUX.synchronize do
-      @parsers ||= []
-      @parsers << object_responding_to_new
-    end
-  end
+  PARSERS = [
+    FormatParser::Parsers::AudioParser,
+    FormatParser::Parsers::DocumentParser,
+    FormatParser::Parsers::ImageParser,
+    FormatParser::Parsers::VideoParser
+  ].freeze
 
   def self.parse_http(url)
     parse(RemoteIO.new(url))
@@ -22,34 +16,31 @@ module FormatParser
 
   def self.parse(io)
     io = Care::IOWrapper.new(io) unless io.is_a?(Care::IOWrapper)
+    limited_io = ReadLimiter.new(io, max_bytes: 512 * 1024, max_reads: 64 * 1024, max_seeks: 64 * 1024)
 
     # Always instantiate parsers fresh for each input, since they might
     # contain instance variables which otherwise would have to be reset
     # between invocations, and would complicate threading situations
-    parsers = @parsers.map(&:new)
-
-    parsers.each do |parser|
-      # We need to rewind for each parser, anew
-      io.seek(0)
-      # Limit how many operations the parser can perform
-      limited_io = ReadLimiter.new(io, max_bytes: 512*1024, max_reads: 64*1024, max_seeks: 64*1024)
+    results = PARSERS.inject({}) do |memo, parser|
       begin
-        if info = parser.information_from_io(limited_io)
-          return info
-        end
+        key = parser.name.split('::').last.downcase.gsub('parser', '').to_sym
+        io.seek(0)
+        memo[key] = parser.new.call(limited_io)
+        memo
       rescue IOUtils::InvalidRead
         # There was not enough data for this parser to work on,
         # and it triggered an error
+        memo[key] = nil
+        memo
       rescue ReadLimiter::BudgetExceeded
         # The parser tried to read too much - most likely the file structure
         # caused the parser to go off-track. Strictly speaking we should log this
         # and examine the file more closely.
+        memo[key] = nil
+        memo
       end
     end
-    nil # Nothing matched
-  end
-
-  Dir.glob(__dir__ + '/parsers/*.rb').sort.each do |parser_file|
-    require parser_file
+    return nil if results.values.all?(&:nil?) # Nothing matched
+    FormatParser::Result.new(results)
   end
 end
