+++ lib/video.rb
@@ -1,0 +1,5 @@
+module FormatParser
+  class Video
+    # TODO: Figure what's the metadata worthy to be presented about videos.
+  end
+end
