+++ spec/format_parser_spec.rb
@@ -11,6 +11,24 @@ describe FormatParser do
     expect(FormatParser.parse(d)).to be_nil
   end
 
+  context 'when two natures are returned' do
+    # TODO: Add FactoryGurl to the specs.
+    let(:gif) { FormatParser::Image.new(format: :gif) }
+    let(:audio) { FormatParser::Audio.new(format: :aiff) }
+    before do
+      expect_any_instance_of(FormatParser::Parsers::Image::GIFParser).to receive(:call).and_return(gif)
+      expect_any_instance_of(FormatParser::Parsers::Audio::AIFFParser).to receive(:call).and_return(audio)
+    end
+    subject { FormatParser.parse(StringIO.new('')) }
+
+    it { expect(subject.natures).to include(:audio) }
+    it { expect(subject.natures).to include(:image) }
+    it { expect(subject.audio).to eq(audio) }
+    it { expect(subject.image).to eq(gif) }
+    it { expect(subject.document).to be_nil }
+    it { expect(subject.video).to be_nil }
+  end
+
   describe 'with fuzzing' do
     it "returns either a valid result or a nil for all fuzzed inputs at seed #{RSpec.configuration.seed}" do
       r = Random.new(RSpec.configuration.seed)
