+++ spec/parsers/exif_parser_spec.rb
@@ -1,7 +1,6 @@
 require 'spec_helper'
 
-describe FormatParser::EXIFParser do
-  
+describe FormatParser::Parsers::Image::EXIFParser do
   # ORIENTATIONS = [
   #   :top_left,
   #   :top_right,
@@ -17,13 +16,13 @@ describe FormatParser::EXIFParser do
     Dir.glob(fixtures_dir + '/exif-orientation-testimages/jpg/*.jpg').each do |jpeg_path|
       filename = File.basename(jpeg_path)
       it "is able to parse #{filename}" do
-        parser = FormatParser::EXIFParser.new(:jpeg, File.open(jpeg_path, 'rb'))
+        parser = described_class.new(:jpeg, File.open(jpeg_path, 'rb'))
         parser.scan_image_exif
         expect(parser).not_to be_nil
 
         expect(parser.orientation).to be_kind_of(Symbol)
         # Filenames in this dir correspond with the orientation of the file
-        expect(filename.include?(parser.orientation.to_s)).to be true 
+        expect(filename.include?(parser.orientation.to_s)).to be true
       end
     end
   end
@@ -32,13 +31,13 @@ describe FormatParser::EXIFParser do
     Dir.glob(fixtures_dir + '/exif-orientation-testimages/tiff-*/*.tif').each do |tiff_path|
       filename = File.basename(tiff_path)
       it "is able to parse #{filename}" do
-        parser = FormatParser::EXIFParser.new(:tiff, File.open(tiff_path, 'rb'))
+        parser = described_class.new(:tiff, File.open(tiff_path, 'rb'))
         parser.scan_image_exif
         expect(parser).not_to be_nil
 
         expect(parser.orientation).to be_kind_of(Symbol)
         # Filenames in this dir correspond with the orientation of the file
-        expect(filename.include?(parser.orientation.to_s)).to be true 
+        expect(filename.include?(parser.orientation.to_s)).to be true
       end
     end
   end
