+++ spec/parsers/jpeg_parser_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-describe FormatParser::JPEGParser do
+describe FormatParser::Parsers::Image::JPEGParser do
   describe 'is able to parse all the examples from FastImage' do
     Dir.glob(fixtures_dir + '/*.jpg').each do |jpeg_path|
       it "is able to parse #{File.basename(jpeg_path)}" do
-        parsed = subject.information_from_io(File.open(jpeg_path, 'rb'))
+        parsed = described_class.new.call(File.open(jpeg_path, 'rb'))
         expect(parsed).not_to be_nil
-        expect(parsed.file_nature).to eq(:image)
-        expect(parsed.file_type).to eq(:jpg)
+        expect(parsed.nature).to eq(:image)
+        expect(parsed.format).to eq(:jpg)
 
         expect(parsed.width_px).to be_kind_of(Integer)
         expect(parsed.width_px).to be > 0
@@ -21,7 +21,7 @@ describe FormatParser::JPEGParser do
   describe 'is able to parse all the JPEG exif examples from FastImage' do
     Dir.glob(fixtures_dir + '/exif-orientation-testimages/jpg/*.jpg').each do |jpeg_path|
       it "is able to parse #{File.basename(jpeg_path)}" do
-        parsed = subject.information_from_io(File.open(jpeg_path, 'rb'))
+        parsed = described_class.new.call(File.open(jpeg_path, 'rb'))
         expect(parsed).not_to be_nil
 
         expect(parsed.orientation).to be_kind_of(Symbol)
@@ -32,5 +32,4 @@ describe FormatParser::JPEGParser do
       end
     end
   end
-
 end
