+++ spec/parsers/tiff_parser_spec.rb
@@ -1,14 +1,14 @@
 require 'spec_helper'
 
-describe FormatParser::TIFFParser do
+describe FormatParser::Parsers::Image::TIFFParser do
   describe 'is able to parse all the examples from FastImage' do
     Dir.glob(fixtures_dir + '/TIFF/*.tif').each do |tiff_path|
       it "is able to parse #{File.basename(tiff_path)}" do
-        parsed = subject.information_from_io(File.open(tiff_path, 'rb'))
+        parsed = subject.call(File.open(tiff_path, 'rb'))
 
         expect(parsed).not_to be_nil
-        expect(parsed.file_nature).to eq(:image)
-        expect(parsed.file_type).to eq(:tif)
+        expect(parsed.nature).to eq(:image)
+        expect(parsed.format).to eq(:tif)
 
         expect(parsed.width_px).to be_kind_of(Integer)
         expect(parsed.width_px).to be > 0
@@ -22,7 +22,7 @@ describe FormatParser::TIFFParser do
   describe 'is able to parse all the TIFF exif examples from FastImage' do
     Dir.glob(fixtures_dir + '/exif-orientation-testimages/tiff-*/*.tif').each do |tiff_path|
       it "is able to parse #{File.basename(tiff_path)}" do
-        parsed = subject.information_from_io(File.open(tiff_path, 'rb'))
+        parsed = subject.call(File.open(tiff_path, 'rb'))
         expect(parsed).not_to be_nil
 
         expect(parsed.orientation).to be_kind_of(Symbol)
@@ -33,5 +33,4 @@ describe FormatParser::TIFFParser do
       end
     end
   end
-
 end
